+++ packages/nft-checkout/src/operations/evm/evm-operation.ts
@@ -1,4 +1,5 @@
 import { extend, ref, toRaw } from '@distributedlab/reactivity'
+import { BN } from '@distributedlab/tools'
 import type { Token } from '@rarimo/bridge'
 import type { IProvider } from '@rarimo/provider'
 import {
@@ -31,8 +32,6 @@ import {
   bnFromAmountLike,
   checkout as _checkout,
   estimate,
-  getAmountWithBridgeFee,
-  getAmountWithoutBridgeFee,
   getEstimation,
   getPaymentTokensWithPairs,
   getSameChainSwapToToken,
@@ -41,6 +40,8 @@ import {
   handleCorrectProviderChain,
   isSameChainOperation,
 } from './helpers'
+
+const RARIMO_BRIDGE_FEE = 2.5
 
 /**
  * An operation on an EVM chain.
@@ -252,7 +253,9 @@ export const EVMOperation = (provider: IProvider): CheckoutOperation => {
       chainIdTo,
       from,
       to,
-      amountOut: getAmountWithBridgeFee(price),
+      amountOut: Amount.fromBN(
+        bnFromAmountLike(price).addPercent(RARIMO_BRIDGE_FEE),
+      ),
       slippage,
     }
 
@@ -267,11 +270,22 @@ export const EVMOperation = (provider: IProvider): CheckoutOperation => {
       estimationWithFee.amountIn,
     ).toDecimals(swapToToken.decimals)
 
+    // Amount from which percent will be subtracted on the backend side has
+    // precision chainFromUSDCAmountOut.decimals, thus we need to cut off
+    // the extra precision from the chainFromUSDCAmountOut.raw, so we will create
+    // a new BN instance from the chainFromUSDCAmountOut.value
+    const amountIn = Amount.fromBN(
+      BN.fromBigInt(
+        chainFromUSDCAmountOut.value,
+        chainFromUSDCAmountOut.decimals,
+      )
+        .subPercent(RARIMO_BRIDGE_FEE)
+        .toDecimals(estimationWithFee.amountIn.decimals),
+    )
+
     intermediateOpts = {
       ...estimationWithFee,
-      amountIn: getAmountWithoutBridgeFee(
-        chainFromUSDCAmountOut.toDecimals(estimationWithFee.amountIn.decimals),
-      ),
+      amountIn,
       amountOut: price,
     }