+++ packages/shared/src/const/chains.ts
@@ -1,7 +1,7 @@
-import { NEAR_CHAIN_IDS } from '@/const/near'
 import { ChainTypes } from '@/enums'
 
 import { EVM_CHAIN_IDS } from './evm'
+import { NEAR_CHAIN_IDS } from './near'
 import { SOLANA_CHAIN_IDS } from './solana'
 
 export const CHAIN_IDS = {