+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-rc.7] - 2023-07-17
+### Changed
+- `all` - Updated `@distributedlab` packages to the latest versions (`1.0.0-rc.1`)
+
+## [2.0.0-rc.6] - 2023-07-14
+### Changed
+- `all` - Updated `@distributedlab` packages to the latest versions (`1.0.0-rc.0`)
+
 ## [2.0.0-rc.5] - 2023-07-13
 ### Fixed
 - `@rarimo/identity-gen-iden3` - fix global config
@@ -365,7 +373,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.5...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.7...HEAD
+[2.0.0-rc.7]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.6...2.0.0-rc.7
+[2.0.0-rc.6]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.5...2.0.0-rc.6
 [2.0.0-rc.5]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.4...2.0.0-rc.5
 [2.0.0-rc.4]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.3...2.0.0-rc.4
 [2.0.0-rc.3]: https://github.com/rarimo/js-sdk/compare/2.0.0-rc.2...2.0.0-rc.3